Use Firebase hosting's preview feature for per-PR deploying (#36)
* Use Firebase hosting's preview feature.

Firebase offers a preview feature where a version of the site is
deployed to a "preview" channel. With this change, a preview channel is
created with each PR. A link is also posted to the PR allowing easing
browsing.

This is a work in progress. I'm force submitting now to see if there's any
change on the github end that occurs after submit. Right now my working
theory is that github is just not propagating the token.