Prettify the project. (#59)
Ran the following command:

> npx prettier --write

I had to manually remove css dependencies (like boostrap) which we
should prettify. This feels like there's an issue with our source tree
since I should be able to do something like the below and have it do the
right thing:

> npx prettier --write src/*

That is, to say that our tree should clearly distinguish between
dependencies and source code.