Adds Firestore Excercise.
1. Added a dependency on the `esm` module based on this article:

https://alxgbsn.co.uk/2019/02/22/testing-native-es-modules-mocha-esm/

Mocha interprets all files as CommonJS but the ES6 syntax is a bit
better to dev on. Since this test code doesn't run in a browser we don't
need to transpile it to CommonJS before shiping. Note: This is an area
of a lot of personal confusion despite reading much of the literature.
I think the consequence of this is that for testing js that runs in the
browswer we'll definitely need to stick to ES6 (and transpile to
a widely availabe ES5 before shipping). This all implies webpack or
Babel although I don't understand the implications fully.

2. All implemenations must export a Run() method which takes a db.
3. All implemenations must update the testSuites variable to register
   there implemenation.