Extracts common nav bar into a custom HTML element. (#98)
A new HTML Element was created using ShadowDom. 
This new element named (navBar) contains a basic version of our navBar. 
This means it has no scroll wheel effect and it is not sticky. 
This new element is in all pages except Demo.html and Index.html and requires a  navBar.js import which cannot be deferred.

How It Works: 

1. The fetch method retrieves all the code from navBar.html and then runs the defineNavBar function
2. The defineNavBar function creates a new class that contains a constructor that activates the ShadowDom and connects it to the navBar.html.
3.  Then Custom Elements creates the new Tag and and appends the class NavBar to our new Tag.

Because of this, whenever you use the tag <navBar> in a HTML page, it will run the code from navBar.html.
 

* Revised NavBar

* Implementation B

* Add NavBar Template To All Pages

* Added Coach Lilly's Changes